Add new spell checker based on vim.fn.spellbadword()
Whilst removing dependency on the FFI (which does not work on windows).
This should also be more performant since most lines will require only 1
call to `vim.fn.spellbadword()`.

The FFI version has been kept as an alternate spellchecker as an optional
fallback in case the new vim.fn imp has bugs.
