+++ README.md
@@ -59,5 +59,11 @@ the default settings:
 require('spellsitter').setup {
   hl = 'SpellBad',
   captures = {'comment'},  -- set to {} to spellcheck everything
+
+  -- Spellchecker to use. values:
+  -- * vimfn: built-in spell checker using vim.fn.spellbadword()
+  -- * ffi: built-in spell checker using the FFI to access the
+  --   internal spell_check() function
+  spellchecker = 'vimfn',
 }
 ```
